feat: column support for `array_append`, `array_prepend`, `array_position` and `array_positions` (#6805)
* test: sqllogictests with columns for array_append, array_prepend, array_position and array_positions

* feat: column support for array_append and array_prepend

* feat: column support for array_position and array_positions

* fix: error type

+++ datafusion/expr/src/built_in_function.rs
@@ -488,9 +488,9 @@ impl BuiltinScalarFunction {
             )))),
             BuiltinScalarFunction::ArrayLength => Ok(UInt8),
             BuiltinScalarFunction::ArrayNdims => Ok(UInt8),
-            BuiltinScalarFunction::ArrayPosition => Ok(UInt8),
+            BuiltinScalarFunction::ArrayPosition => Ok(UInt64),
             BuiltinScalarFunction::ArrayPositions => {
-                Ok(List(Arc::new(Field::new("item", UInt8, true))))
+                Ok(List(Arc::new(Field::new("item", UInt64, true))))
             }
             BuiltinScalarFunction::ArrayPrepend => Ok(List(Arc::new(Field::new(
                 "item",
